# Reuse thrift pageheader buffers