--- a/file.go
+++ b/file.go
@@ -525,8 +525,10 @@ func (f *filePages) ReadPage() (Page, error) {
 		return nil, io.EOF
 	}
 
+	header := getPageHeader()
+	defer putPageHeader(header)
+
 	for {
-		header := new(format.PageHeader)
 		if err := f.decoder.Decode(header); err != nil {
 			return nil, err
 		}
@@ -588,7 +590,9 @@ func (f *filePages) readDictionary() error {
 	defer putBufioReader(rbuf, pool)
 
 	decoder := thrift.NewDecoder(f.protocol.NewReader(rbuf))
-	header := new(format.PageHeader)
+
+	header := getPageHeader()
+	defer putPageHeader(header)
 
 	if err := decoder.Decode(header); err != nil {
 		return err
@@ -759,4 +763,18 @@ func getBufioReaderPool(size int) *sync.Pool {
 	pool := &sync.Pool{}
 	bufioReaderPool[size] = pool
 	return pool
+}
+
+var pageHeaderPool = &sync.Pool{}
+
+func getPageHeader() *format.PageHeader {
+	h, _ := pageHeaderPool.Get().(*format.PageHeader)
+	if h != nil {
+		return h
+	}
+	return new(format.PageHeader)
+}
+
+func putPageHeader(h *format.PageHeader) {
+	pageHeaderPool.Put(h)
 }